Temporarily move When and Then structs into Core namespace.
